Initial commit of deodex functionality in dexlib and baksmali
git-svn-id: https://smali.googlecode.com/svn/trunk@435 55b6fa8a-2a1e-11de-a435-ffa8d773f76a

+++ baksmali/src/main/java/org/jf/baksmali/Adaptors/Format/Instruction22sMethodItem.java
@@ -42,6 +42,6 @@ public class Instruction22sMethodItem extends InstructionFormatMethodItem<Instru
     protected void setAttributes(StringTemplate template) {
         template.setAttribute("RegisterA", formatRegister(instruction.getRegisterA()));
         template.setAttribute("RegisterB", formatRegister(instruction.getRegisterB()));
-        template.setAttribute("Literal", instruction.getLiteral());
+        template.setAttribute("Literal", (int)instruction.getLiteral());
     }
 }
